+++ R/compile_pseudo_pop.R
@@ -20,8 +20,6 @@
 #' `seq(min(w)+delta_n/2,max(w), by=delta_n)`.
 #' @param nthread An integer value that represents the number of threads to be
 #' used by internal packages.
-#' @param optimized_compile If TRUE, uses counts to keep track of number of
-#' replicated pseudo population.
 #' @param ... Additional parameters.
 #'
 #' @note
@@ -56,7 +54,6 @@
 #'                          gps_model = "parametric",
 #'                          bin_seq = NULL,
 #'                          nthread = 1,
-#'                          optimized_compile=TRUE,
 #'                          matching_fun = "matching_l1",
 #'                          covar_bl_method = 'absolute',
 #'                          covar_bl_trs = 0.1,
@@ -66,12 +63,10 @@
 #'
 compile_pseudo_pop <- function(data_obj, ci_appr, gps_model,
                                bin_seq, nthread,
-                               optimized_compile, ...) {
+                               ...) {
 
   # Checking arguments
-  check_args_compile_pseudo_pop(ci_appr = ci_appr,
-                                optimized_compile = optimized_compile,
-                                ...)
+  check_args_compile_pseudo_pop(ci_appr = ci_appr, ...)
 
   if (!(is.object(data_obj) && !isS4(data_obj))) {
     stop("data_obj should be a S3 object.")
@@ -89,7 +84,6 @@ compile_pseudo_pop <- function(data_obj, ci_appr, gps_model,
                                      bin_seq,
                                      gps_model,
                                      nthread,
-                                     optimized_compile,
                                      ...)
       logger::log_info("Finished compiling pseudo population ",
                       " (Pseudo population data size: {nrow(matched_set)})")