+++ NEWS.md
@@ -6,7 +6,7 @@
 
 ### Changed
 
-* 
+* Only optimized compilation is supported. In the previous versions, this approach is known as `optimzied_compile == TRUE`.
 
 ### Added